fix(types): document query argument in graphql interface (#60)
The jsdoc for the `graphql(query, options)` method overload specified the
first argument as a route, when it's really a GraphQL query. This
updates the documentation to reflect that.

This also adds a line to the `graphql(options)` method overload that
specifies that the GraphQL query must be specified in the `options`
parameter.